refactor(livechat): convert `packages/livechat/src/lib/locale.js`
